requirements.txt
Flask>=2.0.0
python-docx>=1.0.0
openai>=1.0.0
python-dotenv>=0.19.0 # Added for .env support
httpx>=0.25.0 # Added for explicit proxy handling
